requirements.txt
aiohttp==3.9.1
aiohttp-retry==2.8.3
aiosignal==1.3.1
anyio==3.7.1
async-timeout==4.0.3
attrs==23.1.0
cachetools==5.3.2
certifi==2023.11.17
charset-normalizer==3.3.2
click==8.1.7
exceptiongroup==1.2.0
fastapi==0.105.0
frozenlist==1.4.1
google-api-core==2.15.0
google-auth==2.25.2
google-cloud-dialogflow-cx==1.30.0
googleapis-common-protos==1.62.0
grpcio==1.60.0
grpcio-status==1.60.0
h11==0.14.0
idna==3.6
iniconfig==2.0.0
multidict==6.0.4
packaging==23.2
pluggy==1.3.0
proto-plus==1.23.0
protobuf==4.25.1
pyasn1==0.5.1
pyasn1-modules==0.3.0
pydantic==1.10.13
PyJWT==2.8.0
pytest==7.4.3
python-multipart==0.0.6
requests==2.31.0
rsa==4.9
sniffio==1.3.0
starlette==0.27.0
twilio==8.11.0
typing_extensions==4.9.0
urllib3==2.1.0
uvicorn==0.25.0
yarl==1.9.4